[add] change int64 to uint

+++ go/pkg/infra/entity/article.go
@@ -10,8 +10,8 @@ const (
 )
 
 type Article struct {
-	ID        int64     `gorm:"column:id"`
-	UserID    int64     `gorm:"column:user_id"`
+	ID        uint      `gorm:"column:id"`
+	UserID    uint      `gorm:"column:user_id"`
 	Title     string    `gorm:"column:title"`
 	Body      string    `gorm:"column:body"`
 	CreatedAt time.Time `gorm:"column:created_at"`
