+++ go/pkg/infra/entity/user.go
@@ -14,7 +14,7 @@ const (
 )
 
 type User struct {
-	ID          int64     `gorm:"column:id"`
+	ID          uint      `gorm:"column:id"`
 	Name        string    `gorm:"column:name"`
 	Description string    `gorm:"column:description"`
 	Email       string    `gorm:"column:email"`
